Index of raytracer() now 16 bit to save memory

+++ tracer.py
@@ -69,7 +69,7 @@ def __cyl_raytracer(p1, p2, mesh, ncell=0):
         return None, None
     hindex1 = np.searchsorted(mesh.jbins, h1, side="right")-1
     hindex2 = np.searchsorted(mesh.jbins, h2)-1
-    index = np.zeros((4, hindex2-hindex1+1), dtype=int)
+    index = np.zeros((4, hindex2-hindex1+1), dtype="int16")
     index[0,:] = rindex
     index[1,:] = [i for i in range(hindex1, hindex2+1)]
     index[2,:] = tindex
@@ -151,7 +151,7 @@ def __xyz_raytracer(p1, p2, mesh, ncell=0):
     cdists = np.diff(cdists[first:last], prepend=0)
     cells = cells[first:last]
     # print(cells)
-    index = np.zeros((4, cells.shape[0]), dtype=int)
+    index = np.zeros((4, cells.shape[0]), dtype="int16")
     index[0:3] = np.transpose(cells)
     index[3,:] = ncell
     return index, cdists
